requirements.txt
google-api-python-client
pandas
numpy
matplotlib
seaborn
scikit-learn
xgboost
isodate
python-dotenv
flask
joblib
emoji
